# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SaaS Vertex AI Gateway — API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/video:
    post:
      operationId: generateVideo
      tags: [ai-proxy]
      summary: Generate a video with Veo (async by default, optional synchronous wait)
      description: |
        Starts a Veo video generation job. Returns 202 with a jobId that can be polled
        via `/v1/video/{jobId}/status`. If `wait=true` is passed, the server polls
        internally and returns 200 with the final `videoUrl` (up to 4 minutes).

        Duplicate requests (same apiKey + model + prompt + duration) within 10 minutes
        return the original jobId with a `duplicateOf` field and are NOT re-billed.
      parameters:
        - name: wait
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: If true, server polls internally and returns the completed video in one call.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoRequest"
      responses:
        "200":
          description: |
            Returned in two cases:
              1. `wait=true` and the job finished (status=completed or failed).
              2. Duplicate request within the 10-minute idempotency window — the
                 original jobId is echoed with `duplicateOf` and `note` fields
                 (status may still be `pending`; no new charge).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoJobResponse"
        "202":
          description: Video generation started (async) or wait-mode timed out before completion.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoJobResponse"
  /v1/video/{jobId}/status:
    get:
      operationId: getVideoStatus
      tags: [ai-proxy]
      summary: Poll video generation job status
      description: |
        Returns the current job status. Uses lightweight API-key authentication
        (validates key only; works even when the account balance reaches zero,
        so already-paid jobs can still be retrieved). Job ownership is enforced
        via the originating API key.
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoJobResponse"
  /v1/video/{jobId}/download:
    get:
      operationId: downloadVideo
      tags: [ai-proxy]
      summary: Download a completed video as a real video/mp4 file
      description: |
        Returns the finished video as a real `video/mp4` binary (with a `Content-Disposition`
        attachment header) so that HTTP clients (n8n, Zapier, browsers) can save or forward
        it without any base64 decoding. Handles three upstream URI shapes:
        `data:video/mp4;base64,…` (decoded inline), `gs://…` (proxied via the
        Google-authenticated Storage API), and `https://…` (proxied only for
        allow-listed Google Storage hosts — SSRF guarded). Uses lightweight API-key
        auth so zero-balance accounts can still download their paid jobs.
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Binary MP4 video file
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "409":
          description: Job is still processing — poll the status endpoint first
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    VideoRequest:
      type: object
      properties:
        model:
          type: string
          description: "Video model ID (e.g. veo-3.1-generate-001)"
          default: "veo-3.1-generate-001"
        prompt:
          type: string
        durationSeconds:
          description: Length of the generated clip in seconds. Accepts a string for n8n/Zapier compatibility.
          oneOf:
            - type: integer
            - type: string
          default: 5
        sampleCount:
          description: Number of video samples to generate. Accepts a string for n8n/Zapier compatibility.
          oneOf:
            - type: integer
            - type: string
          default: 1
        wait:
          type: boolean
          description: If true, the server polls internally and returns the completed video in a single call.
          default: false
        waitTimeoutMs:
          type: integer
          description: Optional override (ms) for the synchronous wait timeout. Capped at 240000.
          default: 180000
      required: [prompt]
    VideoJobResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          description: "One of: pending, completed, failed"
        videoUrl:
          type: ["string", "null"]
          description: "Final URL when status=completed. May be a data URL (data:video/mp4;base64,…), a gs:// URI, or an https URL. Use /v1/video/{jobId}/download to fetch the raw MP4 bytes."
        errorMessage:
          type: ["string", "null"]
        model:
          type: string
        costUsd:
          type: ["number", "null"]
        duplicateOf:
          type: ["string", "null"]
          description: Present when the request was deduplicated against a prior jobId (no new charge).
        note:
          type: ["string", "null"]
          description: Human-readable note, e.g. "Duplicate request detected — returning existing job (no additional charge)."
        statusUrl:
          type: ["string", "null"]
          description: Full URL to poll job status.
        pollIntervalSeconds:
          type: ["integer", "null"]
          description: Suggested polling interval in seconds.
        estimatedSeconds:
          type: ["integer", "null"]
          description: Rough time-to-completion estimate.
      required: [jobId, status, model]
...